requirements.txt
fastapi
groq
uvicorn
phidata